Create Farmland_Distance_To_Water.js
This Google Earth Engine (GEE) script calculates the Euclidean distance from the centroid of each agricultural land parcel to the nearest water body. It leverages the ESRI 2022 Land Cover data to identify water features and exports the results, including the calculated distance in meters, for each farmland parcel as a CSV file to Google Drive.